Move piexif import to server file for code splitting
