+++ src/utility/exif.ts
@@ -1,6 +1,5 @@
 import { OrientationTypes, type ExifData } from 'ts-exif-parser';
 import { formatNumberToFraction, roundToString } from './number';
-import * as PiExif from 'piexifjs';
 
 const OFFSET_REGEX = /[+-]\d\d:\d\d/;
 
@@ -56,22 +55,4 @@ export const formatExposureCompensation = (exposureCompensation?: number) => {
   } else {
     return undefined;
   }
-};
-
-export const removeGpsFromFile = async (
-  fileBytes: ArrayBuffer
-): Promise<Blob> => {
-  const base64 = Buffer.from(fileBytes).toString('base64');
-  const base64Url = `data:image/jpeg;base64,${base64}`;
-
-  const exifObject = PiExif.load(base64Url) as Record<string, any>;
-  delete exifObject.GPS;
-  const exifDataWithoutGps = PiExif.dump(exifObject);
-
-  const data = PiExif.insert(
-    exifDataWithoutGps,
-    base64Url,
-  );
-
-  return fetch(data, { cache: 'no-store' }).then(res => res.blob());
 };
